Move SplineInterplator into separate file and update
