+++ src/corecel/grid/Interpolator.hh
@@ -12,8 +12,6 @@
 #include "corecel/Macros.hh"
 #include "corecel/Types.hh"
 #include "corecel/cont/Array.hh"
-#include "corecel/math/Algorithms.hh"
-#include "corecel/math/PolyEvaluator.hh"
 
 #include "detail/InterpolatorTraits.hh"
 
@@ -86,53 +84,6 @@ class Interpolator
     using YTraits_t = detail::InterpolatorTraits<YI, real_type>;
 };
 
-//---------------------------------------------------------------------------//
-/*!
- * Interpolate using a cubic spline.
- *
- * Given a set of \f$ n \f$ data points \f$ (x_i, y_i) \f$ such that \f$ x_0 <
- * x_1 < \dots < x_{n - 1} \f$, a cubic spline \f$ S(x) \f$ interpolating on
- * the points is a piecewise polynomial function consisting of \f$ n - 1 \f$
- * cubic polynomials \f$ S_i \f$ defined on \f$ [x_i, x_{i + 1}] \f$. The \f$
- * S_i \f$ are joined at \f$ x_i \f$ such that both the first and second
- * derivatives, \f$ S'_i \f$ and \f$ S''_i \f$, are continuous.
- *
- * The \f$ i^{\text{th}} \f$ piecewise polynomial \f$ S_i \f$ is given by:
- * \f[
-   S_i(x) = a_0 + a_1(x - x_i) + a_2(x - x_i)^2 + a_3(x - x_i)^3,
- * \f]
- * where \f$ a_i \f$ are the polynomial coefficients, expressed in terms of the
- * second derivatives as:
- * \f{align}{
-   a_0 &= y_i \\
-   a_1 &= \frac{\Delta y_i}{\Delta x_i} - \frac{\Delta x_i}{6} \left[ S''_{i +
-          1} + 2 S''_{i} \right] \\
-   a_2 &= \frac{S''_i}{2} \\
-   a_3 &= \frac{1}{6 \Delta x_i} \left[ S''_{i + 1} - S''_i \right]
- * \f}
- */
-template<typename T = ::celeritas::real_type>
-class SplineInterpolator
-{
-  public:
-    //!@{
-    //! \name Type aliases
-    using real_type = T;
-    using Point = Array<T, 3>;
-    //!@}
-
-  public:
-    // Construct with left and right values for x and y
-    inline CELER_FUNCTION SplineInterpolator(Point left, Point right);
-
-    // Interpolate
-    inline CELER_FUNCTION real_type operator()(real_type x) const;
-
-  private:
-    T x_lower_;  //!< Lower grid point \f$ x_i \f$
-    Array<T, 4> a_;  //!< Spline coefficients
-};
-
 //! Linear interpolation
 template<class T>
 using LinearInterpolator = Interpolator<Interp::linear, Interp::linear, T>;
@@ -170,31 +121,6 @@ CELER_FUNCTION Interpolator<XI, YI, T>::Interpolator(Point left, Point right)
 
 //---------------------------------------------------------------------------//
 /*!
- * Construct with left and right values for x and y.
- */
-template<class T>
-CELER_FUNCTION
-SplineInterpolator<T>::SplineInterpolator(Point left, Point right)
-{
-    enum
-    {
-        X = 0,
-        Y = 1,
-        Y_DD = 2  //!< Second derivative
-    };
-
-    CELER_EXPECT(left[X] < right[X]);
-
-    x_lower_ = left[X];
-    T h = right[X] - left[X];
-    a_[0] = left[Y];
-    a_[1] = (right[Y] - left[Y]) / h - h / 6 * (right[Y_DD] + 2 * left[Y_DD]);
-    a_[2] = T(0.5) * left[Y_DD];
-    a_[3] = 1 / (6 * h) * (right[Y_DD] - left[Y_DD]);
-}
-
-//---------------------------------------------------------------------------//
-/*!
  * Interpolate linearly on the transformed type.
  */
 template<Interp XI, Interp YI, class T>
@@ -207,17 +133,6 @@ Interpolator<XI, YI, T>::operator()(real_type x) const -> real_type
 
     CELER_ENSURE(!std::isnan(result));
     return result;
-}
-
-//---------------------------------------------------------------------------//
-/*!
- * Interpolate using a cubic spline.
- */
-template<class T>
-CELER_FUNCTION auto SplineInterpolator<T>::operator()(real_type x) const
-    -> real_type
-{
-    return PolyEvaluator{a_}(x - x_lower_);
 }
 
 //---------------------------------------------------------------------------//
